further simplify import func
